Moved model.dart from models/data => models
