+++ lib/src/models/data/serial.dart
@@ -2,7 +2,7 @@ import "package:rover_dashboard/data.dart";
 import "package:rover_dashboard/models.dart";
 import "package:rover_dashboard/services.dart";
 
-import "model.dart";
+import "../model.dart";
 
 /// Connects to and monitors Serial devices connected over USB.
 ///
