+++ lib/src/models/data/video.dart
@@ -4,7 +4,7 @@ import "package:rover_dashboard/data.dart";
 import "package:rover_dashboard/models.dart";
 import "package:rover_dashboard/services.dart";
 
-import "model.dart";
+import "../model.dart";
 
 /// A data model to stream video from the rover.
 class VideoModel extends Model {
