+++ lib/src/models/model.dart
